requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.48
aiosqlite==0.20.0
greenlet==3.3.2

# Data validation
pydantic==2.12.5
pydantic[email]==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.7.1

# Type hints support
typing-extensions==4.15.0

# Security
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
